# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tu Plan Saludable API",
    "description": "API completa para la gestión de planes de dieta, entrenamiento y seguimiento de progreso. Permite a entrenadores crear y gestionar planes personalizados para sus clientes, así como el seguimiento del progreso y comunicación entre usuarios.",
    "version": "1.0.0",
    "contact": {
      "name": "Tu Plan Saludable",
      "email": "support@tuplansaludable.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.tuplansaludable.com",
      "description": "Servidor de producción"
    }
  ],
  "paths": {
...
     "/api/clients": {
       "get": {
         "tags": ["Clients"],
         "summary": "Obtener lista de clientes",
         "description": "Obtener lista paginada de clientes del entrenador autenticado",
         "security": [
           {
             "bearerAuth": []
           }
         ],
         "parameters": [
           {
             "name": "page",
             "in": "query",
             "description": "Número de página",
             "required": false,
             "schema": {
               "type": "number",
               "default": 1,
               "minimum": 1
             }
           },
           {
             "name": "limit",
             "in": "query",
             "description": "Número de elementos por página",
             "required": false,
             "schema": {
               "type": "number",
               "default": 10,
               "minimum": 1,
               "maximum": 50
             }
           },
           {
             "name": "search",
             "in": "query",
             "description": "Término de búsqueda",
             "required": false,
             "schema": {
               "type": "string"
             }
           },
           {
             "name": "sortBy",
             "in": "query",
             "description": "Campo para ordenar",
             "required": false,
             "schema": {
               "type": "string",
               "enum": ["createdAt", "firstName", "lastName", "email"],
               "default": "createdAt"
             }
           },
           {
             "name": "sortOrder",
             "in": "query",
             "description": "Orden de clasificación",
             "required": false,
             "schema": {
               "type": "string",
               "enum": ["asc", "desc"],
               "default": "desc"
             }
           }
         ],
         "responses": {
           "200": {
             "description": "Lista de clientes obtenida exitosamente",
             "content": {
               "application/json": {
                 "schema": {
                   "allOf": [
                     {
                       "$ref": "#/components/schemas/ApiResponse"
                     },
                     {
                       "type": "object",
                       "properties": {
                         "data": {
                           "type": "array",
                           "items": {
                             "$ref": "#/components/schemas/User"
                           }
                         },
                         "pagination": {
                           "type": "object",
                           "properties": {
                             "page": {
                               "type": "number"
                             },
                             "limit": {
                               "type": "number"
                             },
                             "total": {
                               "type": "number"
                             },
                             "pages": {
                               "type": "number"
                             }
                           }
                         }
                       }
                     }
                   ]
                 }
               }
             }
           },
           "401": {
             "description": "No autenticado",
             "content": {
               "application/json": {
                 "schema": {
                   "$ref": "#/components/schemas/ApiResponse"
                 }
               }
             }
           },
           "403": {
             "description": "No autorizado (solo entrenadores)",
             "content": {
               "application/json": {
                 "schema": {
                   "$ref": "#/components/schemas/ApiResponse"
                 }
               }
             }
           }
         }
       }
     },
...
    "/api/upload/progress": {
      "post": {
        "tags": ["Upload"],
        "summary": "Subir imágenes de progreso",
        "description": "Subir múltiples imágenes de progreso para el usuario autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["progressImages"],
                "properties": {
                  "progressImages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Imágenes de progreso (JPG, PNG, máximo 5 archivos, 5MB cada uno)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imágenes subidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Imágenes subidas exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "urls": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uri"
                          },
                          "description": "URLs de las imágenes subidas"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Archivos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "413": {
            "description": "Archivos demasiado grandes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload/chat": {
      "post": {
        "tags": ["Upload"],
        "summary": "Subir archivos para chat",
        "description": "Subir archivos para usar en el sistema de chat",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["chatFiles"],
                "properties": {
                  "chatFiles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Archivos para chat (imágenes, documentos, máximo 3 archivos, 10MB cada uno)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivos subidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Archivos subidos exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "urls": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uri"
                          },
                          "description": "URLs de los archivos subidos"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Archivos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "413": {
            "description": "Archivos demasiado grandes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido al autenticarse"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectId",
            "description": "ID único del usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del usuario"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario"
          },
          "lastName": {
            "type": "string",
            "description": "Apellido del usuario"
          },
          "role": {
            "type": "string",
            "enum": ["USER", "TRAINER"],
            "description": "Rol del usuario"
          },
          "profilePicture": {
            "type": "string",
            "format": "uri",
            "description": "URL de la imagen de perfil"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "description": "Fecha de nacimiento"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "other"],
            "description": "Género del usuario"
          },
          "height": {
            "type": "number",
            "description": "Altura en centímetros"
          },
          "weight": {
            "type": "number",
            "description": "Peso en kilogramos"
          },
          "activityLevel": {
            "type": "string",
            "enum": ["SEDENTARY", "LIGHTLY_ACTIVE", "MODERATELY_ACTIVE", "VERY_ACTIVE", "EXTREMELY_ACTIVE"],
            "description": "Nivel de actividad física"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["LOSE_WEIGHT", "GAIN_WEIGHT", "MAINTAIN_WEIGHT", "BUILD_MUSCLE", "IMPROVE_FITNESS"]
            },
            "description": "Objetivos del usuario"
          },
          "medicalConditions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Condiciones médicas"
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Alergias del usuario"
          },
          "trainerId": {
            "type": "string",
            "format": "objectId",
            "description": "ID del entrenador asignado"
          },
          "clients": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "objectId"
            },
            "description": "IDs de los clientes (solo para entrenadores)"
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del usuario"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "clientCount": {
            "type": "number",
            "description": "Número de clientes (solo para entrenadores)"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
} 